Refactored old code with leak references, updated all tests, deleted obsolete files
