+++ Tests/all_tests.py
@@ -1,0 +1,3 @@
+exec(open('emission_tests.py', 'r').read())
+exec(open('ldar_tests.py', 'r').read())
+exec(open('results_tests.py', 'r').read())
