+++ feast/DetectionModules/__init__.py
@@ -1,7 +1,4 @@
-from . import null
 from . import abstract_detection_method
-from . import tech_detect
-from . import tiered_detect
 from . import repair
 from . import comp_survey
 from . import site_survey
