+++ feast/EmissionSimModules/__init__.py
@@ -1,0 +1,5 @@
+# indicating the EmissionSimModules is a package
+from . import emission_class_functions
+from . import simulation_classes
+from . import simulation_functions
+from . import infrastructure_classes
