--- feast/GeneralClassesFunctions/__init__.py
@@ -1,7 +1,0 @@
-# indicating the GeneralClassesFunctions is a package
-from . import emission_class_functions
-from . import plotting_functions
-from . import results_analysis_functions
-from . import simulation_classes
-from . import simulation_functions
-from. import site_emission_methods
