+++ feast/ResultsProcessing/__init__.py
@@ -1,0 +1,2 @@
+from . import plotting_functions
+from . import results_analysis_functions
