+++ feast/__init__.py
@@ -1,5 +1,5 @@
 # indicating that FEAST is a package
-from . import GeneralClassesFunctions
+from . import EmissionSimModules
 from . import DetectionModules
 from . import field_simulation
 from . import InputData
